requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
cookiecutter==2.6.0
dj-database-url==2.1.0
Django==5.0.6
django-browser-reload==1.12.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cloudinary-storage==0.3.0
django-js-asset==2.2.0
django-tailwind==3.8.0
django-tinymce==4.0.0
django-widget-tweaks==1.5.0
gunicorn==22.0.0
idna==3.7
Jinja2==3.1.4
mailjet-rest==1.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
